aispirelabs_backend/requirements.txt

Django==5.1.0
djangorestframework==3.15.2
django-cors-headers==4.4.0
PyJWT==2.9.0
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.8.3
cryptography==43.0.1
